+++ src/main/java/com/project/traplaner/member/controller/MemberController.java
@@ -1,21 +1,15 @@
 package com.project.traplaner.member.controller;
 
 import com.project.traplaner.entity.Member;
-import com.project.traplaner.entity.Travel;
 import com.project.traplaner.member.service.MemberService;
 import com.project.traplaner.member.dto.SignUpRequestDto;
-import com.project.traplaner.mypage.dto.response.TravelListResponseDTO;
-import com.project.traplaner.mypage.service.MyPageBoardService;
 import com.project.traplaner.util.FileUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @Controller
 @RequestMapping("/members")
@@ -96,39 +90,6 @@ public class MemberController {
         System.out.println("[dbg] naver-sign-in 진입!!!");
         log.info("naver-sing-in 진입");
 
-    }
-
-    private final MyPageBoardService myPageBoardService;
-
-    @GetMapping("/my-page")
-    public String myPage(){
-
-        return "member/my-page";
-    }
-
-    @GetMapping("/my-page/my-board")
-    public String myBoard() {
-
-
-
-        return "member/my-board";
-    }
-
-
-    // 마이페이지 나의 여행
-    @GetMapping("/my-page/my-plan")
-    public String myPlan(
-             Model model) {
-
-        int memberId = 2;
-        System.out.println("asdasdasdasdas");
-        List<TravelListResponseDTO> dtoList = myPageBoardService.getList(memberId);
-
-        log.info("dtoList: {}", dtoList);
-
-        model.addAttribute("list", dtoList);
-
-        return "member/my-plan";
     }
