+++ src/main/webapp/WEB-INF/views/member/my-board.jsp
@@ -79,7 +79,7 @@
 </style>
 
 <div class="container">
-  <h1 id="mypage_h1">프로필 관리</h1>
+  <h1 id="mypage_h1">내 게시물</h1>
   <div class="mypage_section">
     <div class="mypage_section1">
       <img
@@ -89,8 +89,9 @@
               alt=""
       />
       <div class="manage_box">
-        <a href="">계정관리</a><a style="font-weight: bold" href="/members/my-page/my-board">내 게시물</a
-      ><a href="/members/my-page/my-plan">나의 여행</a><a href="">좋아요한 게시물</a>
+        <a href="">계정관리</a><a style="font-weight: bold" href="/my-page/mytravelboard">내 게시물</a
+      ><a  href="/my-page/mytravel">나의 여행</a>
+        <a href="/my-page">여행일정</a><a href="/mypage/favorite">좋아요한 게시물</a>
       </div>
     </div>
     <div class="mypage_section2">
